+++ frontend/src/views/PaymentView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Payment from '../components/Payment.vue'
+import Payment from '../components/PaymentForm.vue'
 </script>
 
 <template>
